fix(autocomplete): avoid duplicate params when org matches query
